# docs/v1.swagger.yaml
...
swagger: '2.0'
info:
  title: SIARHES API
  description: API para acesso aos dados do SIARHES.
  version: "1.0.0"
host: 127.0.0.1:3000
basePath: /v1
schemes:
  - http
  - https
produces:
  - application/json
security:
  - clientCredentials:
    - siarhes_basico
    - siarhes_admin

# Definições dos Endpoints
paths:
...
  /organograma:
    get:
      operationId: listaOrganograma
      description: |
        Retorna as informações sobre o organograma. <br />
        Ordenação: *nome
      parameters:
        - name: empresa
          in: query
          description: Código da Empresa
          required: false
          type: integer
        - name: subempresa
          in: query
          description: Código da SubEmpresa
          required: false
          type: integer
      responses:
        200:
          description: Lista de setores.
          schema:
            $ref: '#/definitions/Setor'
        400:
          $ref: '#/responses/Error'
        401:
          $ref: '#/responses/AuthError'
  /organograma/setores:
    get:
      operationId: listaSetores
      description: Retorna as informações sobre os setores de uma empresa.
      parameters:
        - name: empresa
          in: query
          description: Código da Empresa
          required: true
          type: integer
        - name: setor
          in: query
          description: Código do Setor
          required: true
          type: string
        - name: subempresa
          in: query
          description: |
            Código da SubEmpresa <br />
            Se for preenchido, a busca de setores acima se limitará apenas a setores da mesma subempresa.
          required: false
          type: integer
        - name: nivel
          in: query
          description: |
            Nível Hierárquico Para Prequisa <br />
            "vazio" - Apenas o setor informado. <br />
            "acima" - Setores acima do informado. <br />
            "abaixo" - Setores abaixo do informado.
          required: false
          type: string
      responses:
        200:
          description: Lista de setores.
          schema:
            $ref: '#/definitions/Setor'
        400:
          $ref: '#/responses/Error'
        401:
          $ref: '#/responses/AuthError'
  /organograma/setor/servidores:
    get:
      operationId: listaServidoresSetor
      description: Retorna quais servidores estão em um determinado setor.
      parameters:
        - name: empresa
          in: query
          description: Código da Empresa
          required: true
          type: integer
        - name: setor
          in: query
          description: Código do Setor
          required: true
          type: string
      responses:
        200:
          description: Lista de servidores.
          schema:
            properties:
              numfunc:
                type: integer
              numvinc:
                type: integer
              nome:
                type: string
              gestor:
                type: string
                description: S/N - indica se o servidor é o gestor do setor.
        400:
          $ref: '#/responses/Error'
        401:
          $ref: '#/responses/AuthError'
  /organograma/setor/gestor:
    get:
      operationId: getGestorSetor
      description: |
        Retorna qual o gestor de um determinado setor.
        Se não houver responsável, busca o responsável pelo setor acima.
      parameters:
        - name: empresa
          in: query
          description: Código da Empresa
          required: true
          type: integer
        - name: setor
          in: query
          description: Código do Setor
          required: true
          type: string
      responses:
        200:
          description: Informações do gestor.
          schema:
            properties:
              numfunc:
                type: integer
              numvinc:
                type: integer
              nome:
                type: string
              setor:
                type: string
                description: Setor do gestor.
              nomesetor:
                type: string
                description: Descrição do setor.
              setor_origem:
                type: string
                description: Apenas preenchido quando o gestor não for do próprio setor pesquisado.
        400:
          $ref: '#/responses/Error'
        401:
          $ref: '#/responses/AuthError'
...
definitions:
  Error:
    type: object
    properties:
      message:
        type: string
...
  Setor:
    properties:
      EMPRESA:
        type: integer
      SUBEMPRESA:
        type: integer
      SETOR:
        type: string
      NOMESETOR:
        type: string
      NOMESETORLONGO:
        type: string
      PAISETOR:
        type: string
      DATAINI:
        type: string
      DATAFIM:
        type: string
      TIPOSETOR:
        type: string
      LOGRADOURO:
        type: string
      NUMENDER:
        type: integer
      COMPLEMENTO:
        type: string
      BAIRRO:
        type: string
      CEP:
        type: string
      MUNICIPIO:
        type: string
      FONE:
        type: string

# Definições de Responses
responses:
  Error:
    description: Erro em caso de parâmetro inválido.
    schema:
      $ref: '#/definitions/Error'
  AuthError:
    description: Erro de autenticação.
    schema:
      $ref: '#/definitions/Error'

# Definições de Perfis de Segurança
securityDefinitions:
  clientCredentials:
    description: |
      O token deve ser obtido atráves do Acesso Cidadão, através da URL https://acessocidadao.es.gov.br/is/connect/token. <br /> <br />
      Para obter o token deve ser passado os headers a seguir. <br />
      <strong>grant_type</strong> -> client_credentials <br />
      <strong>client_id</strong> -> Seu identificador de client. <br />
      <strong>client_secret</strong> -> Seu client secret. <br />
      <strong>scope</strong> -> siarhes_basico <br /><br />
      Para fazer as chamadas na API, basta incluir o header abaixo. <br />
      <strong>Authorization</strong> -> Bearer <token>. <br /> <br />
      O client tem acesso ter apenas na empresa/subempresa cadastrada nos claims.<br />
      Atenção! Este tipo de autorização não deve ser utilizado, em ambientes onde não se pode guardar de forma segura as credenciais do client. Ex App Móvel ou Cliente Web<br /> <br />
    type: oauth2
    flow: application
    tokenUrl: https://acessocidadao.es.gov.br/is/connect/token
    scopes:
      siarhes_basico: Permite o acesso de leitura a todos os endpoints da API.